BlogProject/requirements.txt
# general modules
Django==5.2.7
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
djangorestframework
setuptools

# third party modules
drf-yasg[validation]
django-filter
django-debug-toolbar==6.1.0
django-extensions==4.1
django-robots==6.1
markdown
coreapi==2.3.3
djangorestframework-simplejwt
django-mail-templated
djoser
# development module

pytest
pytest-django
flake8
black
isort
